requirements.txt
cobra>=0.21
lxml == 4.6.2
pyparsing == 2.4.7
memote == 0.12.0
requests == 2.25.1
json2html == 1.3.0
beautifulsoup4 == 4.9.3
PyYAML == 5.4.1
